Custom Authorization

+++ Restaurants.Application/Dishes/Commands/DeleteDishes/DeleteDishesCommandHandler.cs
@@ -1,17 +1,25 @@
 ﻿using MediatR;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDishes;
 
-public class DeleteDishesCommandHandler(IRestaurantsRepository restaurantsRepository, IDishesRepository dishesRepository) : IRequestHandler<DeleteDishesCommand>
+public class DeleteDishesCommandHandler(IRestaurantsRepository restaurantsRepository,
+    IDishesRepository dishesRepository,
+    IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteDishesCommand>
 {
     public async Task Handle(DeleteDishesCommand request, CancellationToken cancellationToken)
     {
         var restaurant = await restaurantsRepository.GetByIdAsync(request.RestaurantId);
         if (restaurant is null)
             throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
+
+        if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Update))
+            throw new ForbiddenAccessException();
+
         await dishesRepository.Delete(restaurant.Dishes);
     }
 }
